requirements.txt.txt
numpy
matplotlib
pandas
tensorflow
scikit-learn
pyserial
